use namespace dcx for the common funcions in the WORDS_BIGENDIAN and WORDS_NEED_ALIGNMENT case
The functions are declared in namespace dcx in common/include/interp.h,
but were defined in namespace dsx in similar/3d/interp.cpp. This broke
building on the RPi. Since all these functions are indentical in d1x and
d2x, this fix uses namespace dcx also for the definition.

Fixes: 32051298ae3f5a1ce5cae21377b2a7f99d9a649f