Add getaddrinfo SCons option to optionally build without getaddrinfo()/freeaddrinfo()
This is a backport of kp's "Add configure test and fallback code to
support ancient Windows" commit (20e72128de76c230920c029b169e617866224129)
to pre-unification code base. However, contrary to kp's version, this does
add an option instead of an automated test. It also sets the _WIN32_WINNT
version only locally in net_udp.c, and only if getaddrinfo() is to be used.